Fix go module paths for v2 modules
This ensures that the kubeclarity components can be used in go.mod files
without falling back to the "incompatible" v0.0.0-<timestamp>-<commit>
format and the need for a replace statement.

Along with this change, a new process needs to take place before a
release of kubeclarity is going to be made. Before tag a commit which
bumps all the go.mod files in kubeclarity to the version number of the
release must be merged to ensure that the inter-component dependecies
work when a component is included in a go.mod outside of this repo.

When the release is made then there needs to be a tag for each of the
components in a sub-directory e.g. api/v2.7.1 to allow go mod to find
the module within the repo at the right version.