Add support for Trivy as an SBOM analyzer
This the first part of an effort to introduce Trivy support to
kubeclarity.

This commit adds Trivy as an SBOM analyzer which can be used in
conjunction with the other SBOM analyzers or by itself to scan the
contents of container images or directories and produce a merged result.

Along with Trivy we introduce a new input type "rootfs" which allows
kubeclarity to scan operating system root file systems like unpacked
containers or virtual machine volumes.