kubeclarity: Use input SBOM format directly as the input for grype
Previously we always converted to syft-json before giving it to grpye
and because of a bug in the syft conversion logic we were using it
resulted in a broken input for certain packages.

This commit updates the scanner logic so that which ever SBOM format we
receive from the user we pass it directly to grpye for scanning. This
does mean that if we are given a broken syft-json input then it will
still error, and we have a ticket open in the syft project to resolve
this.