Remove cyclonedx formatter double buffer
The cyclonedx formatter is wrapping the bytes.Buffer in bufio.NewWriter
which results in a buffer wrapping a buffer, this causes issues when
Flush isn't called on the bufio writer as not all the data is saved to
the bytes.Buffer.

This causes issues when used for json marshaling resulting in no bytes
getting output.

bytes.Buffer already implements the io.Writer interface as long as you
use the pointer so we do not need to wrap it in the extra buffer.