kubeclarity: Output cyclonedx JSON SBOM by default
Grype has an issue loading Cyclonedx XML when the list of components in
the SBOM is empty. This can occur when scanning a scratch container as
there are no dependencies included.

To workaround this issue this commit changes the default SBOM format
that the kubeclarity analyzer output to Cyclonedx JSON instead of XML
ensuring that the output of kubeclarity analyze can be safely passed to
kubeclarity scan.

To force Cyclonedx XML the environment variable ANALYZER_OUTPUT_FORMAT
can be set to "cyclonedx".