kubeclarity: Formalise analyzer results return value as cyclonedx BOM struct
This commit standardise the result format for analyzers to cyclonedx.
Each analyzer is then responsible for converting their implementations
output into a cdx.BOM before returning it for merging. This removes a
lot of encoding and decoding as well as some dangerous casting which
could break. The ANALYZER_OUTPUT_FORMAT environment variable now only
controls the output format *after* merge.