kubeclarity: Remove redundant and dead code
This commit removes a number of functions that have become redundant due
to previous changes, most notably the updated converter package which
can handle more types on input and output. The formatter package is
removed completely as the converter package now does everything the
formatter package used to be needed for.