Improve Docker build performance when re-running builds
Moving the docker ARG directives to where they are required prevents
docker re-evaulating all the steps in the Dockerfile when the args
change.

Adding Dockerfile buildkit caches to the go build step prevents the need
for go to recompile everything on every build, and also removes the need
to put separate layers for `go mod download`.
https://github.com/moby/buildkit/blob/master/frontend/dockerfile/docs/reference.md#run---mount

Without this PR:
`make docker` on a fresh system (docker system prune -a): 38 mins
`make docker` again with a small edit in shared/: 35 mins

With this PR:
`make docker` on a fresh system: 14mins
`make docker` again with a small edit in shared/: 1 min 30 seconds