Indicate doneness by closing scan done channel
Switch to a safer pattern of creating and closing a channel per scan to
indicate doneness instead of using messages on the channel which can
be lost if the scan sends a message before the caller is ready to
receive it.

Adds checks to ensure that calling scan/start multiple times so
that API users will receive a 409 instead of 500 error.

Also ensures that if a scan is aborted by a user, when returning to the
UI its possible to see that the scan was aborted.

Fixes #283