vmclarity: Remove go.mod from subpackages except API and fix Makefile
This commit removes the go.mod from the subpackages runtime_scan and
backend as they import each other in multiple places resulting in
managing the same set of dependecies in three different places.

Once these go.mods are removed then the Makefile targets for linting,
fixing and testing can be cleaned up significantly and all packages
treated as one module.

This commit also fixes an issue in the docker targets which results in
the cli and backend containers getting tagged with the same image name
"vmclarity" when they should be "vmclarity-cli" and "vmclarity-backend".

One final change is that the Dockerfiles have been updated to use the
buildkit caches for the go compilations which speed us local docker
build times significantly.