kubeclarity: Understand more severity strings and convert them to our model
Cyclonedx and a number of other tools have severity levels which we
didn't understand, this results in log messages about unknown severity
types. This commit adds the severity levels "info", "none" and "unknown"
as aliases of the kubeclarity negligible severity level.

It also makes the conversion common between the runtime_k8s_scanner and
cli so that we only need to manage the severity strings in one place.