vmclarity: Fix private network communication in VMClarity VPC
Public subnets require an elastic IP address to talk to the internet
even if we don't require inbound connections.

This commit adds a private subnet for the Scanner VMs with a NAT gateway
to provide internet connectivity and a Scanner specific SecurityGroup.

The SecurityGroup allows us to explictly allow only API communication to
the VMClarity backend as well as SSH access from the VMClarity backend
to the scanner for debugging.