vmclarity: Fix indentation scanner config in Scanner VM cloud-init userdata
This ensures that when we include the rendered scanner configuration
yaml in the cloud-init template, it is indented correctly so that it
doesn't break the yaml formatting of the rest of the template.

This commit adds the sprig library to enhance the functions available to
the golang templating library, then uses the indent function from that
to fix the indentation of the yaml.