vmclarity: Have longer timeout for copied snapshot than initial snapshot creation
A snapshot that is copied between regions takes longer to reach ready
state than the snapshots which are created in the same region. This
commit moves the timeout control for this actions to the job manager
instead of the provider client so that we can have different timeouts
for different behaviours.