vmclarity: Disable additional properties for majority of schemas in API
By disabling additional properties in the API it ensures that if someone
sends a property that we don't understand the API server responds with a
400 error.

This is particularly useful if you make a spelling mistake when crafting
an API request because the mispelled field will result in an error
instead of being accepted and ignored as an addtional property.

The only schemas which can't have additionalProperties disabled are
common schemas which are included in "allOf" statements, that add
additional fields by combining schemas. Instead additionalProperties: false
is added to the combined schema instead.