vmclarity: Include cloud formation lint in validation checks
This commit adds cfn linting to our lint target in the Makefile so that
when validating the code locally or in CI, the VmClarity.cfn is also
checked not just the go code.

This requires cfn-lint to be installed by the developer as there are
different options for how to install this based on the developer's
setup. A link to the cfn-lint documentation is provided in the Makefile.

In CI https://github.com/marketplace/actions/cfn-lint-action is used to
install cfn-lint prior to calling `make check`.