Add separate CLI option for waiting for server attached status
This commit adds a new CLI option which tells the CLI to wait for the
server attached status.

Previously it did wait as part of the mount-attached-volume option, but
that meant it was not possible to mount attached volumes without the
server being involved, for example if running on a system where the
volume is already mounted.

Now you can specify just the --mount-attached-volume option to mount
existing volumes, or both options to wait for the server to attach and
then do the mount.