Make AWS Scanner VM InstanceType configurable
This exposes a new environment variable configuration AWS_INSTANCE_TYPE
for the orchestrator which configures the instance type to use for the
scanner VMs. It defaults to t2.large.

A new parameter has been added to the cloud formation to allow user's to
choose which instance type they wish to use at install time.