vmclarity: Update VMClarity init so stack upgrades apply correctly
This change tweaks the init configs for VMClarity server so that cfn-hup
and cfn-init hooks restart the services correctly when the cloud
formation stack updates the VMClarity server configuration.

Without this change the cfn-hup and cfn-init notice the change and
update the systemd unit files, but the services don't get restarted so
they continue to run the old version.

This change also lowers the cfn-hup polling time from 15 minutes to 5
minutes, so that the VMClarityServer notices that the cloudformation has
been updated sooner.