Fix issue with double counting in findings trends graph (#307)
This commit fixes an issue where the queries building the data for the
findings trends graph accidentily counted some findings twice by
filtering on a range instead of the specific point in time shown on the
graph.

The implementation (backend and UI) for the widget has been changed to
generate and handle a list of specific times between the start and
endtimes requested instead of small ranges, and then calculate the
findings count for that specific point in time.