Remove abused allOf from OpenAPI v3 spec (vmclarity#319)
This commit removes the allOf from several objects in the openapi spec
where properties of those fields needed to be overridden. Instead those
fields are replicated between the object schemas.

Unfortunately this does increase the maintainance requirement as schemas
like ScanConfig and ScanConfigRelationship need to be kept in sync, but
this ensures that the code generation understands the field typing
correctly so that it can be used in the code properly.