vmclarity: Add support for postgres database (vmclarity#299)
This commit add better support to odatasql for SQL variants, as well as
updating the gorm database driver to connect to postgres when configured
to do so.

The cloudformation installer has been updated with a new database
configuration section which allows the user to choose the DB driver
(postgres or sqlite), and also configure options to connect to an
external database.