Add revision number to VMClarity objects and support for If-Match header (vmclarity#269)
* Add revision number to VMClarity objects and support for If-Match header

This adds a revision number to every VMClarity API object, this number
can be used as an eTag in the API through the If-Match header to ensure
that we avoid "lost update" issues while reconciling objects.

* Refactor revision management to reduce duplication

Adds new helper functions which handle checking the revision against
IfMatch and also bumps the revision or sets the initial value if its not
set.