vmclarity: Bump minimum go version to 1.20 (vmclarity#331)
This bumps the minimum go version for VMClarity to 1.20. There are a lot
of useful additions in 1.20 such as the strings.Cut and the ability to
wrap multiple errors from the errors package.

As part of moving to 1.20 we also need to upgrade golangci-lint to 1.52
as the older version does not seem to work with the newer golang code.