Add ability to configure timeout for Scans (vmclarity#273)
* Add ability to configure timeout for Scans

Through the scan configuration a maximum number of seconds
(timeoutSeconds) can be configured.

When a scan is triggered from a configuration this number of seconds
will be added to the start time and saved as timeoutTime on the Scan
object.

If while the Scan is running the current time exceeds the timeoutTime
then the scan is automatically failed and any remaining AssetScans will
be terminated.

If no timeoutTime is provided on a scan it will be automatically timed
out 48 hours after the start time of the scan.

* fix: timeoutSeconds in OData schema

* fix: validation of timeoutSeconds parameter

* refactor: checking if scan is timed out

---------

Co-authored-by: Krisztian Gacsal <chrisgacsal@users.noreply.github.com>