kubeclarity: Add min value validation on Max Scan Parallelism field (#443)
* Add min value validation on Max Scan Parallelism field (fixes #441)

Initially the UI allowed negative entries to be entered in the Max Scan Parallelism (under the scan configuration). This is not valid since negative numbers for parallelism (e.g. Max Scan Parallelism = -20) do not make sense and are not valid according to the API spec. After this change, UI will allow only non-negative entries to be entered. This fixes #441.