vmclarity: Installer for VMClarity on Azure (vmclarity#359)
Introduces an installer for VMClarity on Azure using Bicep / Azure Resource
Manager. VMClarity will be installed into a new Resource Group with the control
plane services running in a single VM. A dedicated VMClarity vNet will be
created with 2 subnets, one for the control plane services which allows
external access to the VMClarity server, and one for the Scanners which is only
allows ingress from the control plane network so that it can be used as a
bastion.

Two roles are added:

* One for discovery and snapshotting within a subscription.

* One for creating resources within the VMClarity resource group for scanning
  (VM, Disks, Network Interfaces)

To install on Azure you can either:

* Use the az CLI to directly deploy the bicep (vmclarity.bicep)

* Manually upload the ARM template (vmclarity.json) to the portal as a template
  spec

* Click the "Deploy to Azure" button in the README to deploy the ARM template
  with the custom UI wizard (this is the recommended method)