vmclarity: Add granular permissions to GitHub action workflows (vmclarity#485)
This commit adds granular permissions to the workflows that need them to
perform their tasks.

The release action needs access to the "contents" of a repo in order to
create a release, and requires access to packages in order to publish
the containers.

The main-merge action only needs access to packages to publish the
"latest" build from the main branch.