vmclarity: Add support for any, all and not ODATA operators (vmclarity#327)
The any() lamdba operator can be used like:

    GET api/Targets?$filter=findingInfo/tags/any(t: t/Key eq 'Name' and t/Value eq 'vmclarity-server')

This will return all Targets with at least one tag which has Key Name
and the Value vmclarity-server.

The all lamba operator can be used like:

    GET api/Targets?$filter=findingInfo/securityGroups/all(g: g/id eq 'sg1' or g/id eq 'sg2')

This will return Targets which are only in the security groups sg1 and
sg2, no other security groups.

The not operator can be used like:

    GET api/Targets?$filter=not(startswith(location, 'us-east-2'))

This will return all Targets which do not have a location which starts
with us-east-2.