vmclarity: feat: introduce helm chart for kubernetes installation (vmclarity#633)
* feat: introduce helm chart for kubernetes installation

This commit adds a helm chart for installing the VMClarity control plane on
Kubernetes. Communications between the control plane services is
performed internal to the cluster. Postgresql is used as the database
which allows the API Server, UI and UI Backend services to be scaled
horizontally.

While the control plane runs on Kubernetes the provider is still
responsible for where the scanner instances run, and how they
communicate back to the control plane.

* Add component specific _helper.tpl and reduced duplication

* Add auxiliary services to helm chart

* Add logLevel configuration for api, grype, orchestrator and uibackend

* Pin nginx to 1.25.1

* Lock down container security context where possible

* Add provider configurations

* Make it possible to configure resource limits and requests

* Add more helm-doc comments

* Add helm docs verification to ci workflow

* Code review comments

* Add swagger ui to helm chart

* Add service account per microservice

* Add orchestrator and uibackend healthchecks