vmclarity: Add success job to ci.yml which only runs if everything else succeeds (vmclarity#649)
This job can be used as the indicator that every required job succeeded by
the branch protection rules. Adding this means we no longer need to
update the branch protection rules every time we make a change to the
workflow instead it can be completely controlled in the workflow by
managing whether "Success" runs successful.