vmclarity: Buffered Asset Discovery (vmclarity#626)
Modify DiscoverAssets interface to return a channel of assets instead of
a list. This lays the foundation to prevent us from keeping big chunks
of assets in memory.

In a future PR, all the providers should be updated to paginate fetching
assets from their downstream APIs and feed the channel.