Bump golangci-lint to 1.54.2 (#504)
This version contains a number of fixes and ensures that it works for
golang 1.21+.

The depguard and tagalign linters were added since 1.52.x which are
temporarily disabled as former seems to be usefull, but requires a
discussion with the team before enabling it while, the latter does not
seem to provide much of a value besides making the structs with struct
tags a bit more readable/organized.

There are several fixes and nolint comments that were added to address
lint issues identified by the new version.