feat(scanners): Add support to all existing scanners for containers (vmclarity#696)
* feat(scanners): Add support to all existing scanners for containers

This commit adds support to all existing scanners for all families for
the image, docker-archive, oci-archive, and oci-directory input types.

In the scenario where a scanner does not support containers natively a
new library containerrootfs has been added which converts the container
into a rootfs folder.

The containerrootfs library supports a cached mode which indentifies
containers based on the source and cache the result so that multiple
scanners (even concurrent ones) attempting to convert the same container
to a rootfs only perform it once. All these cached folders are cleaned
up by a function called once all the families have completed.

* refactor(scanners): handle cleanup better for container rootfs

* fix(scanners): error handling in rootfs cache cleanup

* fix(scanner): make sure that symlinks are supported for container rootfs