fix: lynis testDB path (vmclarity#886)
The `lynis show dbdir` command returns the path to the test database
directory on the where the output string has a trailing newline
character which is not removed before generating the path for the
`tests.db` file. This results a bogus file path liek
`/usr/local/lynis/db /tests.db` and triggers an error as the `tests.db`
file is not accessable.

This fix make sure that the command output is sanitized by removing
leading/trailing whitespaces and the path is cleaned before using it to
assemble the file path for `tests.db`.

Switch to `filepath` from `path` for working with filesystem paths
as the latter is not meant to be used with OS filesystem paths.

Also rename `testdb` variable to `testDB` in order to make it aligned
with Go naming convention.