feat(k8s): Discover container images and running containers (#711)
* feat(kubernetes): Discover container images and running containers

With this addition the kubernetes provider can now discover container
images and running containers from the connected kubernetes cluster.

A new Daemonset is introduced which provides entrypoints on each
kubernetes node to discover the images and containers running there. The
Kubernetes provider will query each instance of the daemonset to create
full view of all the containers and images that exist on the kubernetes
cluster. The Daemonset supports Docker and ContainerD, other container
runtimes can be added in the future through a defined interface.

* refactor: Docker build for CR Discovery Server

* refactor(helm): update CR Discovery deployment

* update image name for CR Discovery Server
* install CR Discovery Server only if the provider is set to
  `kubernetes`

* test(e2e): get version from env if not empty

* test(e2e): fix docker package alias

* test(e2e): set log level for tests

* test(e2e): set timeout for tests

* test(e2e): add K8s discovery

* fix(k8s): return response body on error

* fix(k8s): ignore image size calculation error

---------

Co-authored-by: Krisztian Gacsal <chrisgacsal@users.noreply.github.com>